requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.1
requests==2.31.0
db-dtypes==1.2.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
